feat: impl cache building
