+++ moat/src/server/mod.rs
@@ -15,14 +15,22 @@
 use std::{net::SocketAddr, time::Duration};
 
 use async_trait::async_trait;
-use clap::Parser;
+use bytes::Bytes;
+use bytesize::ByteSize;
+use clap::{Args, Parser};
+use foyer::{
+    DirectFsDeviceOptions, Engine, HybridCache, LargeEngineOptions, RecoverMode, RuntimeOptions, Throttle,
+    TokioRuntimeOptions,
+};
 use pingora::{prelude::*, proxy::http_proxy_service_with_name, server::configuration::ServerConf};
 use serde::{Deserialize, Serialize};
-use url::Url;
 
 use crate::{
     api::service::ApiService,
-    aws::{AwsSigV4Resigner, AwsSigV4ResignerConfig},
+    aws::{
+        resigner::{AwsSigV4Resigner, AwsSigV4ResignerConfig},
+        s3::S3Config,
+    },
     meta::{
         manager::{Gossip, MetaManager, MetaManagerConfig},
         model::{Identity, Peer},
@@ -61,6 +69,39 @@ impl MoatRequest {
     }
 }
 
+#[derive(Debug, Args, Serialize, Deserialize)]
+pub struct CacheConfig {
+    #[clap(long, default_value = "64MiB")]
+    mem: ByteSize,
+
+    #[clap(long)]
+    dir: Option<String>,
+
+    #[clap(long, default_value = "1GiB", requires = "dir")]
+    disk: ByteSize,
+
+    #[clap(long, default_value = "64MiB")]
+    file_size: ByteSize,
+
+    #[clap(flatten)]
+    throttle: Throttle,
+
+    #[clap(long, default_value_t = 4)]
+    flushers: usize,
+
+    #[clap(long, default_value_t = 2)]
+    reclaimers: usize,
+
+    #[clap(long, default_value = "16MiB")]
+    buffer_pool_size: ByteSize,
+
+    #[clap(long, default_value = "quiet")]
+    recover_mode: RecoverMode,
+
+    #[clap(long, default_value_t = 4)]
+    recover_concurrency: usize,
+}
+
 #[derive(Debug, Parser, Serialize, Deserialize)]
 pub struct MoatConfig {
     #[clap(long, default_value = "127.0.0.1:23456")]
@@ -88,25 +129,52 @@ pub struct MoatConfig {
     #[clap(long, default_value_t = 1)]
     weight: usize,
 
-    #[clap(long)]
-    s3_endpoint: Url,
-    #[clap(long)]
-    s3_access_key_id: String,
-    #[clap(long)]
-    s3_secret_access_key: String,
-    #[clap(long)]
-    s3_region: String,
+    #[clap(flatten)]
+    s3_config: S3Config,
+
+    #[clap(flatten)]
+    cache: CacheConfig,
 }
 
 pub struct Moat;
 
 impl Moat {
-    pub fn run(config: MoatConfig) {
+    pub fn run(config: MoatConfig) -> anyhow::Result<()> {
         let runtime = tokio::runtime::Builder::new_multi_thread()
             .enable_all()
             .build()
             .expect("Failed to create Tokio runtime");
         let runtime = Runtime::new(runtime);
+
+        let cache = {
+            let config = config.cache;
+            runtime.block_on(async move {
+                let mut builder = HybridCache::builder()
+                    // TODO(MrCroxx): introduce metrics system.
+                    // .with_metrics_registry(registry)
+                    .memory(config.mem.as_u64() as _)
+                    // TODO(MrCroxx): Count serialized size?
+                    .with_weighter(|path: &String, data: &Bytes| path.len() + data.len())
+                    .storage(Engine::Large(
+                        LargeEngineOptions::new()
+                            .with_flushers(config.flushers)
+                            .with_reclaimers(config.reclaimers)
+                            .with_buffer_pool_size(config.buffer_pool_size.as_u64() as _)
+                            .with_recover_concurrency(config.recover_concurrency),
+                    ))
+                    .with_recover_mode(config.recover_mode)
+                    .with_runtime_options(RuntimeOptions::Unified(TokioRuntimeOptions::default()));
+                if let Some(dir) = config.dir {
+                    builder = builder.with_device_options(
+                        DirectFsDeviceOptions::new(dir)
+                            .with_capacity(config.disk.as_u64() as _)
+                            .with_file_size(config.file_size.as_u64() as _)
+                            .with_throttle(config.throttle),
+                    );
+                }
+                builder.build().await
+            })
+        }?;
 
         let meta_manager = MetaManager::new(MetaManagerConfig {
             identity: config.identity,
@@ -126,10 +194,10 @@ impl Moat {
 
         let api = ApiService::new(meta_manager);
         let resigner = AwsSigV4Resigner::new(AwsSigV4ResignerConfig {
-            endpoint: config.s3_endpoint.clone(),
-            region: config.s3_region.clone(),
-            access_key_id: config.s3_access_key_id.clone(),
-            secret_access_key: config.s3_secret_access_key.clone(),
+            endpoint: config.s3_config.endpoint.clone(),
+            region: config.s3_config.region.clone(),
+            access_key_id: config.s3_config.access_key_id.clone(),
+            secret_access_key: config.s3_config.secret_access_key.clone(),
         });
 
         let mut conf = ServerConf::new().unwrap();
@@ -138,16 +206,17 @@ impl Moat {
         let mut server = Server::new_with_opt_and_conf(None, conf);
         server.bootstrap();
 
-        let s3_host = config.s3_endpoint.host_str().unwrap_or("localhost").to_string();
+        let s3_host = config.s3_config.endpoint.host_str().unwrap_or("localhost").to_string();
         let s3_port = config
-            .s3_endpoint
+            .s3_config
+            .endpoint
             .port()
-            .unwrap_or(if config.s3_endpoint.scheme() == "https" {
+            .unwrap_or(if config.s3_config.endpoint.scheme() == "https" {
                 443
             } else {
                 80
             });
-        let s3_tls = config.s3_endpoint.scheme() == "https";
+        let s3_tls = config.s3_config.endpoint.scheme() == "https";
 
         let proxy = Proxy {
             api,
@@ -155,6 +224,7 @@ impl Moat {
             s3_host,
             s3_port,
             s3_tls,
+            _cache: cache,
         };
         let mut service = http_proxy_service_with_name(&server.configuration, proxy, "moat");
         service.add_tcp(&config.listen.to_string());
@@ -180,6 +250,8 @@ struct Proxy {
     s3_host: String,
     s3_port: u16,
     s3_tls: bool,
+
+    _cache: HybridCache<String, Bytes>,
 }
 
 impl Proxy {
